+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -9,6 +9,9 @@
   - `BootstrapWitness` in `Cardano.Ledger.Keys.Bootstrap`
   - `Addr`, `RewardAccount` in `Cardano.Ledger.Address`
   - `ScriptHash` `ToJSONKey`/`FromJSONKey` in `Cardano.Ledger.Hashes`
+  - `Inclusive` and `Exclusive` in `Cardano.Ledger.BaseTypes`
+  - `DirectDeposits` in `Cardano.Ledger.Address`
+* Add `ToJSONKey`/`FromJSONKey` instances for `AccountId`
 * Change `ChainCode` type to use `ByteArray` instead of `ByteString`
 * Change `bwAttributes` field to use `ByteArray` instead of `ByteString`
 * Add `StAnnTx` type family and `txStAnnTxG` to `Tx` type class
@@ -44,6 +47,7 @@
   - `ScriptHash` `ToJSONKey`/`FromJSONKey` in `Cardano.Ledger.Hashes`
   - `Data` in `Cardano.Ledger.Plutus.Data`
   - `PlutusBinary` in `Cardano.Ledger.Plutus.Language`
+  - `Datum` in `Cardano.Ledger.Plutus.Data` (discriminated schema: `null`/`{"datumhash":...}`/inline data)
 
 ### `cddl`